# src/main/frontend/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /users/{userId}/cart/{cartId}:
    put:
      tags:
      - user-controller
      operationId: updateCartProduct
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: cartId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartProductDto'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CartProduct'
      security:
      - Bearer_Authentication: []
    delete:
      tags:
      - user-controller
      operationId: removeCartProduct
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: cartId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
      security:
      - Bearer_Authentication: []
  /users/{id}:
    get:
      tags:
      - user-controller
      operationId: one
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
      security:
      - Bearer_Authentication: []
    put:
      tags:
      - user-controller
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
      security:
      - Bearer_Authentication: []
    delete:
      tags:
      - user-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
      security:
      - Bearer_Authentication: []
  /users/{id}/addresses/{addressId}:
    put:
      tags:
      - user-controller
      operationId: updateUserAddress
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: addressId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressDto'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AddressDto'
      security:
      - Bearer_Authentication: []
...
  /users/{user_id}/wishlist:
    get:
      tags:
      - user-controller
      operationId: getWishList
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
      security:
      - Bearer_Authentication: []
    post:
      tags:
      - user-controller
      operationId: addToWishList
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: product_id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDto'
      security:
      - Bearer_Authentication: []
    delete:
      tags:
      - user-controller
      operationId: removeFromWishList
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: product_id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
      security:
      - Bearer_Authentication: []
  /users/{id}/cart:
    get:
      tags:
      - user-controller
      operationId: getCartProducts
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartProduct'
      security:
      - Bearer_Authentication: []
    post:
      tags:
      - user-controller
      operationId: addCartProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartProductDto'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CartProduct'
      security:
      - Bearer_Authentication: []
  /users/{id}/addresses:
    post:
      tags:
      - user-controller
      operationId: addUserAddress
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressDto'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AddressDto'
      security:
      - Bearer_Authentication: []
...
  /users:
    get:
      tags:
      - user-controller
      operationId: getUsers
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
      security:
      - Bearer_Authentication: []
  /users/me:
    get:
      tags:
      - user-controller
      operationId: user
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
      security:
      - Bearer_Authentication: []
...
components:
  schemas:
    CartProductDto:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
    CartProduct:
      type: object
      properties:
        id:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        product:
          $ref: '#/components/schemas/Product'
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Product:
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
          format: int64
        description:
          type: string
        price:
          type: number
          format: double
        imageURL:
          type: string
        deleted:
          type: boolean
        category:
          $ref: '#/components/schemas/Category'
    AddressDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        city:
          type: string
        country:
          type: string
        zip:
          type: string
        addressLine:
          type: string
    UserDto:
      required:
      - email
      - firstName
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/AddressDto'
    Address:
      type: object
      properties:
        id:
          type: integer
          format: int64
        city:
          type: string
        country:
          type: string
        zip:
          type: string
        addressLine:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        password:
          type: string
        enabled:
          type: boolean
        cartProducts:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/CartProduct'
    ProductDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        categoryId:
          type: integer
          format: int64
        imageURL:
          type: string
...
  securitySchemes:
    Bearer_Authentication:
      type: http
      scheme: bearer
      bearerFormat: JWT